Refactor: 무한 스크롤 훅 카드 리스트 컴포넌트 전용화
